Add Experience feature (no exceptions no constraints)

+++ app/src/main/java/com/yelysei/hobbyharbor/model/userhobbies/entities/Action.kt
@@ -2,6 +2,6 @@ package com.yelysei.hobbyharbor.model.userhobbies.entities
 
 data class Action (
     val id: Int,
-    val startTime: Int,
-    val endTime: Int
+    val startTime: Long,
+    val endTime: Long
 )
